refactor : add manufacturer in the domain layer
