+++ src/infra/repositories/manufacturer.ts
@@ -1,20 +1,5 @@
+import { Manufacturer, ManufacturerCreateInput } from "../../domain/manufacturer";
 import { prisma } from "../database/prisma";
-
-export type Manufacturer = {
-  manufacturer: ManufacturerCreateInput;
-};
-
-export type ManufacturerCreateInput = {
-  name: string;
-  phone: string;
-  email: string;
-  street: string;
-  number: string;
-  city: string;
-  state: string;
-  zip_code: string;
-  website: string;
-};
 
 export interface IManufacturerRepository {
   create(data: ManufacturerCreateInput): Promise<Manufacturer>;
