+++ src/presentation/controllers/manufacturer.ts
@@ -2,7 +2,7 @@ import { Request } from "express";
 import { IAddManufacturerUseCase } from "../../application/usecases/manufacturer";
 import { Controller, HttpResponse } from "../../domain/controller";
 import { HttpStatusCode } from "../../domain/http";
-import { ManufacturerCreateInput } from "../../infra/repositories/manufacturer";
+import { ManufacturerCreateInput } from "../../domain/manufacturer";
 
 export class AddManufacturerController implements Controller {
   constructor(readonly addManufacturerUseCase: IAddManufacturerUseCase) {}
